Revert "Create module for realm syncing and write unit tests before refactoring"
This reverts commit 693bf39f90014a4116d45abcecaa1c5f2016a335.
